% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app91-package.R
\docType{package}
\name{app91-package}
\alias{app91}
\alias{app91-package}
\title{app91: Understanding the movements in economic growth}
\description{
With this application we aim to provide the user with intuitive understanding of movements in economic growth using interactive visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AyushBipinPatel/app91}
  \item Report bugs at \url{https://github.com/AyushBipinPatel/app91/issues}
}

}
\author{
\strong{Maintainer}: Ayush Patel \email{ayush.ap58@gmail.com}

}
\keyword{internal}
